# Speed up GC.compact