--- a/gc.c
+++ b/gc.c
@@ -8444,7 +8444,7 @@ gc_compact(rb_objspace_t *objspace, int use_toward_empty, int use_double_pages, 
         /* pin objects referenced by maybe pointers */
         garbage_collect(objspace, GPR_DEFAULT_REASON);
         /* compact */
-        gc_compact_after_gc(objspace, use_toward_empty, use_double_pages, TRUE);
+        gc_compact_after_gc(objspace, use_toward_empty, use_double_pages, use_verifier);
     }
     objspace->flags.during_compacting = FALSE;
 }